% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accel_time.R
\name{accel_time}
\alias{accel_time}
\title{Acceleration Over Time}
\usage{
accel_time(max_speed, player_tau, split_time)
}
\arguments{
\item{max_speed}{Max observed speed of the player}

\item{player_tau}{A unique Tau value for each player}

\item{split_time}{Time}
}
\value{
Acceleration reached over time from zero velocity
}
\description{
Acceleration Over Time
}
