% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\alias{optim_player_profile}
\title{Combine Data}
\usage{
combine_data(distance, time, units = "m/s", player_name = NULL)

optim_player_profile(...)
}
\arguments{
\item{distance}{Distance vector in the form of 10, 20, 40 yd or 9.1, 18.3, 36.6 m.}

\item{time}{Split times.}

\item{units}{Typically yards or meters.}

\item{player_name}{Athlete name - optional.}
}
\value{
Builds a player's 40 yard/ 36.6 meter sprint test data frame. This function feeds into combine_profile to model their speed and acceleration over time.
}
\description{
combine_data and combine_profile replace optim_player_profile from the previous update, creating a more consistent coding environment amongst all _data and _profile functions.

Please note: the distance and time vectors comprise of 3 values: time splits at 10, 20, 40 yards/ 9.1, 18.3, 36.6 meters. To visualize how the data set should look, type player_combine into the console.
}
