% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_player_plot.R
\name{combine_player_plot}
\alias{combine_player_plot}
\alias{optim_player_plot}
\title{Combine Player Plot}
\usage{
combine_player_plot(combine_profile, ...)

optim_player_plot(combine_profile, ...)
}
\arguments{
\item{...}{Other combine profiles - optional.}

\item{game_profile}{An athlete's combine profile - required.}
}
\value{
Time vs speed and acceleration for 1+ athletes.
}
\description{
This function returns two curves per athlete: time versus speed and time versus acceleration. The curve moving downward is time versus acceleration, and the curve moving upward is time versus acceleration.

This function replaces previously published functions optim_player_plot and compare_profile_plot. This function also has much simpler syntax, requiring only one argument with an indefinite number of player profiles possible.
}
