% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_profile}
\alias{combine_profile}
\title{Combine Profile}
\usage{
combine_profile(combine_data)
}
\arguments{
\item{combine_data}{Athlete's 40 yd/ 36.6 m data}
}
\value{
Returns a player's theoretical max speed and acceleration values. Tau values are also calculated.

This function works in conjunction with combine_data.

When run, the athlete's theoretical max speed and accel will be printed in the console. If saved as an object and called upon, a formatted data frame will appear. Although it is less aesthetically pleasing, all of the information remains the same.
}
\description{
Combine Profile
}
