% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_player_distance_time_plot.R
\name{compare_player_distance_time_plot}
\alias{compare_player_distance_time_plot}
\title{Distance Time Comparison Plot}
\usage{
compare_player_distance_time_plot(player_profile_1, player_profile_2)
}
\arguments{
\item{player_profile_1}{player 1 profile}

\item{player_profile_2}{player 2 profile}
}
\value{

}
\description{
This function returns a plot comparing two player profiles. These profiles can originate from
two players or can compare a player's modeled and in-situ profiles.
}
\examples{
data(player_a)
data(player_b)
player_a_data <- tracking_data(player_a)
player_a_profile <- player_profile(player_a_data)

player_b_data <- tracking_data(player_b)
player_b_profile <- player_profile(player_b)

player_a_b_compare_plot <- compare_player_plot(player_a_profile, player_b_profile)
}
