% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_player_plot.R
\name{compare_player_plot}
\alias{compare_player_plot}
\title{Player Comparison Plot}
\usage{
compare_player_plot(player_profile_1, player_profile_2)
}
\arguments{
\item{player_profile_2}{}
}
\value{

}
\description{
Player Comparison Plot
}
