% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot1.R
\name{compare_plot1}
\alias{compare_plot1}
\title{Part 1 of Player Comparisons}
\usage{
compare_plot1(player_profile_1)
}
\arguments{
\item{player_profile_1}{}
}
\value{

}
\description{
Part 1 of Player Comparisons
}
