% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_time_to_position.R
\name{compare_time_to_position}
\alias{compare_time_to_position}
\title{Comparing Two Players to Reach a Position}
\usage{
compare_time_to_position(
  player_1_profile,
  player_1_speed,
  player_1_distance,
  player_2_profile,
  player_2_speed,
  player_2_distance
)
}
\arguments{
\item{player_2_profile}{}
}
\value{

}
\description{
Comparing Two Players to Reach a Position
}
