% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_factor.R
\name{conversion_factor}
\alias{conversion_factor}
\title{Conversion Factor}
\usage{
conversion_factor(value, units = "m/s")
}
\arguments{
\item{player_profile}{A data.frame built using the game_profile or optim_profile functions.}
}
\value{
Converts user input units to metric.
}
\description{
Units are extracted from the player's game_data function that are stored in their player profile. Values are returned in their original inputs.
Converting values to metric allows back-end functions to model a player's mechanical sprint qualities in keeping with their intended usage.
}
