% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{distance_speed}
\alias{distance_speed}
\title{Distance Reached at a Given Speed}
\usage{
distance_speed(player_profile, current_speed)
}
\arguments{
\item{player_profile}{Player's f-v-p profile}

\item{current_speed}{Player's current speed}
}
\value{
The distance a player can run to reach a given speed from a velocity >= 0
}
\description{
Distance Reached at a Given Speed
}
