% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_time.R
\name{distance_time}
\alias{distance_time}
\title{Distance Over Time}
\usage{
distance_time(max_speed, player_tau, split_time)
}
\arguments{
\item{max_speed}{Max observed speed of the player}

\item{player_tau}{A unique tau value for each player}

\item{split_time}{Time}
}
\value{
The distance traveled over a given amount of time from zero velocity
}
\description{
Distance Over Time
}
