% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_time_plot.R
\name{distance_time_plot}
\alias{distance_time_plot}
\title{Distance Time Plot}
\usage{
distance_time_plot(player_profile)
}
\arguments{
\item{player_profile}{player profile}
}
\value{
A player's speed and acceleration curves over time
}
\description{
This function returns a plot detailing a player's distance covered over a given amount of time
}
