% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_speed.R
\name{final_speed}
\alias{final_speed}
\title{Final speed reached after a given distance}
\usage{
final_speed(player_profile, current_speed, distance)
}
\arguments{
\item{distance}{}
}
\value{

}
\description{
Final speed reached after a given distance
}
