% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\name{game_data}
\alias{game_data}
\alias{player_data}
\title{Game Data}
\usage{
game_data(speed, acceleration, units = "m/s", player_name = NULL)

player_data(speed, acceleration, units = "m/s", player_name = NULL)
}
\arguments{
\item{speed}{In-game athlete speed vector.}

\item{acceleration}{In-game athlete acceleration vector.}

\item{units}{Units used by the tracking system.}

\item{player_name}{Athlete name - if not provided, an arbitrary player number is assigned.}
}
\value{
Athlete game data for modeling.
}
\description{
This function simplifies large data sets that are provided by positional tracking companies to four variables: speed, acceleration, units, and athlete name. This allows the {midsprint} package to model player profiles while maintaining a small footprint on the computer's memory.

Please note that the units input are those associated with the athlete's speed. If speed and acceleration are recorded in different units, please convert to a consistent format before using this function. You can do this using the conversion_factor function to calculate the conversion constant.
}
