% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_accel_dist_time_plot.R
\name{speed_accel_time_distance_plot}
\alias{speed_accel_time_distance_plot}
\alias{accel_speed_distance_time_plot}
\title{Distance-Time Plot}
\usage{
speed_accel_time_distance_plot(game_data, ...)

accel_speed_distance_time_plot(game_data, ...)
}
\arguments{
\item{game_data}{An athlete's game data - required.}

\item{...}{Other game data - optional.}
}
\value{
A distance-time plot for 1+ athletes.
}
\description{
This plot should be called with the speed_accel_plot, as part of the player_report function to provide context.
}
