% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_player_plot.R
\name{game_player_plot}
\alias{game_player_plot}
\title{Speed and Acceleration Plots over Time}
\usage{
game_player_plot(player_profile)
}
\arguments{
\item{player_profile}{player profile}
}
\value{
A player's speed and acceleration curves over time
}
\description{
This function returns a player's in-situ speed and acceleration v time curves
}
