% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\name{game_profile}
\alias{game_profile}
\alias{player_profile}
\title{Game Profile}
\usage{
game_profile(game_data)

player_profile(game_data)
}
\arguments{
\item{game_data}{An athlete's game data.}
}
\value{
The athlete's game profile.
}
\description{
This function is utilized in conjunction with game_data to build an athlete's in-game profile. When run, player information will be printed in the console. When recalled as an object, information is presented as a data.frame with less formatting.

It is not necessary to run this function when modeling speed-acceleration/ acceleration-speed profiles.
}
