% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_time_distance_plot.R,
%   R/game_time_distance_plot.R
\name{combine_time_distance_plot}
\alias{combine_time_distance_plot}
\alias{game_time_distance_plot}
\title{Combine Time Distance Plot}
\usage{
combine_time_distance_plot(combine_profile, ...)

game_time_distance_plot(game_profile, ...)
}
\arguments{
\item{combine_profile}{An athlete's combine profile - required.}

\item{...}{Other game profiles - optional.}

\item{game_profile}{An athlete's game profile - required.}
}
\value{
Time vs distance for 1+ athletes.

Time vs distance for 1+ athletes.
}
\description{
This plot should accompany game_player_plot or a similar function to provide context.

All values are returned in the original units found in the athlete's profile.
}
