% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_player_plot.R
\name{optim_player_plot}
\alias{optim_player_plot}
\title{Modeled Player Speed Acceleration Curves}
\usage{
optim_player_plot(optim_player_profile)
}
\arguments{
\item{optim_player_profile}{}
}
\description{
This function returns a player's modeled speed and acceleration curves from data collected in events
like the NFL Combine 40 yd dash.
}
