% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_player_profile.R
\name{optim_player_profile}
\alias{optim_player_profile}
\title{Modeled Player Profile}
\usage{
optim_player_profile(distance, splits)
}
\arguments{
\item{distance}{distance traveled}

\item{splits}{split times}
}
\value{
Optimized player profile
}
\description{
This function returns a player profile from events like the NFL Combine 40 yd dash
}
\examples{
player_a_model_profile <- model_player_profile(combine_player_a$dist, combine_player_a$split)
player_a_model_profile
}
