% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_plot.R
\name{player_plot}
\alias{player_plot}
\title{Speed-Acceleration Plot}
\usage{
player_plot(player_profile)
}
\arguments{
\item{player_profile}{}
}
\value{

}
\description{
Speed-Acceleration Plot
}
\examples{
\dontrun {p <- player_plot(player_profile)}
}
