% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_plot_values.R
\name{player_plot_values}
\alias{player_plot_values}
\title{Player Plot Values}
\usage{
player_plot_values(player_profile)
}
\arguments{
\item{player_profile}{A data.frame built using the game_profile or combine_profile functions.}
}
\value{
Speed, acceleration, and distance values in relation to speed.
}
\description{
The player_profile parameter is the player's game or optimal profile. These values are taken from in-game or 40 yd dash data and modeled to represent an athlete's sprint mechanics.

Formulas are based off works by Morin, Samozino, Clark, etc.
This function is not exported.
}
