% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_profile.R
\name{player_profile}
\alias{player_profile}
\title{Player Profile}
\usage{
player_profile(profile_data)
}
\arguments{
\item{profile_data}{Must run profile_data function before player_profile function}
}
\value{
A player's unique profile
}
\description{
Player Profile
}
\examples{
data(player_a)
player_a_data <- tracking_data(player_a$speed, player_a$accel)
player_a_profile <- player_profile(player_a_data)
player_a_profile

}
