% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_data.R
\name{profile_data}
\alias{profile_data}
\title{Profile Data}
\usage{
profile_data(data, speed_column, acceleration_column)
}
\arguments{
\item{acceleration_column}{}
}
\value{
Profiling data from positional tracking data
}
\description{
Profile Data
}
\examples{
\dontrun {tracking_data <- data.frame(speed = rnorm(100, 4, 1),
acceleration = rnorm(100, 2, 0.2))
dat <- profile_data(tracking_data, "speed", "acceleration")
head(dat)}

}
