% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation_distance.R
\name{separation_distance}
\alias{separation_distance}
\title{Title}
\usage{
separation_distance(player_1_profile, player_2_profile, distance_to_travel)
}
\arguments{
\item{dist_of_player_1}{}
}
\value{

}
\description{
Title
}
