% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_accel.R
\name{speed_accel}
\alias{speed_accel}
\title{Speed Acceleration Model Fit}
\usage{
speed_accel(game_data)
}
\arguments{
\item{game_data}{}
}
\value{
Modeled athlete max speed and max accel values, model fit (r_squared), and number of observations kept in the final model output.
}
\description{
This function is based on the **in-situ** speed-acceleration model, originally suggested by Morin and
colleagues. It is currently written as efficiently as possible. Please keep in mind that larger
data set will require more time to process.

The function returns a data.frame containing a player's in-situ max speed and acceleration values,
model fit (r_square), and number of observations (n_observation) kept in the final model output. This
funciton is written to return models with r_square values above 0.95. Well-fit models should return
approximately 45-65 observations.

Processing time: approximately 0.01s per 1000 observations
}
\details{
This function returns a data.frame containing a player's in-situ speed-acceleration values
}
\examples{
data(player_a)
player_a_data ,- tracking_data(player_a$speed, player_a$accel)
player_a_insitu_model <- speed_accel_model(player_a_data)
print(player_a_insitu_model)
}
