% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_accel_observations.R
\name{speed_accel_observations}
\alias{speed_accel_observations}
\alias{accel_speed_observations}
\title{Speed-Acceleration Model Observations}
\usage{
speed_accel_observations(game_data)

accel_speed_observations(game_data)
}
\arguments{
\item{game_data}{An athlete's game data.}
}
\value{
Observations withheld from the game data to build the speed-acceleration model.
}
\description{
This function returns the observations that were utilized to build an athlete's speed-acceleration profile. As suggested by Morin and colleagues, the top two acceleration values per speed bin, cut every 0.2 m/s equal to, or above 3 m/s, are maintained. A linear model is then built and points are removed until an r squared value of 0.95 or greater is reached.

You can also call upon accel_speed_observations for the same output.

WARNING: This function returns the values that make up the final linear model. No intermediary values, the initially binned observations, are returned.

For more information: https://www.researchgate.net/publication/351607405_Individual_acceleration-speed_profile_in-situ_A_proof_of_concept_in_professional_football_players
}
