% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_time.R
\name{speed_time}
\alias{speed_time}
\title{Speed over time}
\usage{
speed_time(max_speed, player_tau, split_time)
}
\arguments{
\item{max_speed}{Max observed speed of the player}

\item{player_tau}{A unique tau value for each player}

\item{split_time}{Time}
}
\value{
Speed the player reaches at a given time from zero velocity
}
\description{
Speed over time
}
