% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{start_report}
\alias{start_report}
\alias{save_report}
\alias{speed_accel_report}
\alias{accel_speed_report}
\alias{game_report}
\alias{combine_report}
\title{Start Report}
\usage{
start_report(report_name)

save_report()

speed_accel_report(game_data, ...)

accel_speed_report(game_data, ...)

game_report(game_profile, ...)

combine_report(combine_profile, ...)
}
\arguments{
\item{report_name}{Name of the report}
}
\value{
A player report for 1+ athletes
}
\description{
The report functions return athlete information on 1+ pages. Each line creates a new page in the .pdf that is saved in your working directory.

Formatting the functions in this manner allows you to create a player report for an individual, a position group, or the entire team with minimal effort.

Please note: you must work sequentially as outlined in the supporting .pdf document. If you do not and you run into errors, try running save_report and start again. It is imperative that save_report be called otherwise the report will not be saved and can interfere with other reports you plan to produce.
}
