% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{tau}
\alias{tau}
\title{Player Tau}
\usage{
tau(profile_data)
}
\arguments{
\item{profile_data}{Must run profile_data function before tau function}
}
\value{
A player's unique tau value
}
\description{
Player Tau
}
