% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_distance.R
\name{time_distance}
\alias{time_distance}
\title{Time Distance}
\usage{
time_distance(player_profile, distance)
}
\arguments{
\item{player_profile}{player profile from player_profile function}

\item{time}{Time}
}
\value{
How long it takes to reach a given distance from zero velocity
}
\description{
Time Distance
}
