% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_speed.R
\name{time_speed}
\alias{time_speed}
\title{Time Speed}
\usage{
time_speed(player_profile, current_speed)
}
\arguments{
\item{player_profile}{player profile from player_profile function}

\item{current_speed}{The player's current speed}
}
\value{
How long it takes to reach max speed with a starting speed >= 0
}
\description{
Time Speed
}
