% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_dist_given_speed.R
\name{time_to_dist_given_speed}
\alias{time_to_dist_given_speed}
\title{Time to Reach a Distance Starting from a Given Speed}
\usage{
time_to_dist_given_speed(player_profile, distance, current_speed)
}
\arguments{
\item{player_profile}{Player's f-v-p profile}

\item{distance}{Distance to travel}

\item{current_speed}{The player's current speed}
}
\value{

}
\description{
Time to Reach a Distance Starting from a Given Speed
}
\examples{
\dontrun {
# requires player_profile in the global environment
# player profile for this example: dat
# see player_profile function for more details
time_given_dist_speed(dat, 40, 3)}

}
