% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_position.R
\name{time_to_position}
\alias{time_to_position}
\title{Time to Position}
\usage{
time_to_position(player_profile, current_speed, distance)
}
\arguments{
\item{distance}{}
}
\value{

}
\description{
Time to Position
}
