% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accel_time.R
\name{accel_time}
\alias{accel_time}
\title{Acceleration Over Time}
\usage{
accel_time(player_profile, split_time)
}
\arguments{
\item{player_profile}{player profile from player_profile function}

\item{split_time}{Time}
}
\value{
Acceleration reached over time from zero velocity
}
\description{
Acceleration Over Time
}
