% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_time.R
\name{distance_time}
\alias{distance_time}
\title{Distance Over Time}
\usage{
distance_time(player_profile, split_time)
}
\arguments{
\item{player_profile}{player profile from player_profile function}

\item{split_time}{Time}
}
\value{
The distance traveled over a given amount of time from zero velocity
}
\description{
Distance Over Time
}
