% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\name{game_data}
\alias{game_data}
\title{Player Game Tracking Data}
\usage{
game_data(speed, acceleration)
}
\arguments{
\item{speed}{speed vector from raw data}

\item{acceleration}{acceleration vector from raw data}
}
\value{
Returns a simplified data frame containing an athlete's speed and acceleration values
}
\description{
game_data() is included in the **midsprint** work flow to provide coaches with a simple method
of cleaning positional tracking data. It also reduces overhead associated with large data sets. Doing
so allows the **midsprint** package to run quickly, smoothly, and efficiently.
}
\details{
This function returns a player's tracking data data frame.
}
\examples{
data(player_a)
player_a_data <- tracking_data(player_a$speed, player_a$accel)
head(player_a_data)
}
