% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_profile.R
\name{game_profile}
\alias{game_profile}
\title{Player Profiling}
\usage{
game_profile(game_data)
}
\arguments{
\item{game_data}{}
}
\value{
A player's unique in-situ player profile
}
\description{
This function returns a data.frame containing a player's in-situ profile. The data.frame
includes:
* max observed speed
* max observed acceleration
* calculated in-situ tau value
}
\details{
This function returns a player's in-situ player profile
}
\examples{
\dontrun data(player_a)
player_a_data <- tracking_data(player_a$speed, player_a$accel)
player_a_profile <- player_profile(athlete_a)
player_a_profile#'
}
