% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_plot_values.R
\name{player_plot_values}
\alias{player_plot_values}
\title{Player Plot Values}
\usage{
player_plot_values(player_profile)
}
\arguments{
\item{player_profile}{Must run player_profile function before player_plot_value function}
}
\value{
Plot values for f-v-p profile
}
\description{
Player Plot Values
}
