% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_profile.R
\name{player_profile}
\alias{player_profile}
\title{Player Profile}
\usage{
player_profile(profile_data)
}
\arguments{
\item{profile_data}{Must run profile_data function before player_profile function}
}
\value{
A player's unique profile
}
\description{
Player Profile
}
\examples{
\dontrun {tracking_data <- data.frame(speed = rnorm(100, 4, 1),,
acceleration = rnorm(100, 2, 0.2))
dat <- profile_data(tracking_data, "speed", "acceleration")
head(dat)
player <- player_profile(dat)
print(player)}

}
