% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_accel.R
\name{speed_accel}
\alias{speed_accel}
\alias{accel_speed}
\title{Speed-Acceleration Model}
\usage{
speed_accel(game_data)

accel_speed(game_data)
}
\arguments{
\item{game_data}{An athlete's game data.}
}
\value{
An athlete's theoretical max speed and max accel.
}
\description{
This function returns an athlete's theoretical maximal speed and acceleration from their game data, called using the game_data function.
When called upon, you will see the athlete's information displayed in the console. When you save the function to an object, it will return a data.frame with the player's information. The data.frame is less formatted than the original call but contains the same information.

The model relies on multiple helper functions. These functions convert the original game data to metric, and the linear model is fit using a least squares regression. The model is fit to a minimum r squared value of 0.95.

The speed-acceleration function is modelled based on Morin and colleagues work: https://www.researchgate.net/publication/351607405_Individual_acceleration-speed_profile_in-situ_A_proof_of_concept_in_professional_football_players

You can also call upon accel_speed for the same output.
}
