% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_accel_plot.R
\name{speed_accel_plot}
\alias{speed_accel_plot}
\title{In-Situ Speed-Acceleration Plot}
\usage{
speed_accel_plot(game_data)
}
\arguments{
\item{game_data}{}
}
\value{
In-situ speed-accel player profile plot
}
\description{
The plot returned includes the following components:
* all data player observations (grey)
* final observations modeled (red)
* regression line (black line)
* legend containing player details (modeled max speed, acceleration, and tau)
}
\details{
This model returns a player's in-situ speed-acceleration plot
}
\examples{
data(player_a)
player_a_data <- tracking_data(player_a$speed, player_a$accel)
player_a_insitu_plot <- speed_accel_plot(player_a_data)
player_a_insitu_plot
}
