% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_time.R
\name{speed_time}
\alias{speed_time}
\title{Speed over time}
\usage{
speed_time(player_profile, split_time)
}
\arguments{
\item{player_profile}{player profile from player_profile function}

\item{split_time}{Time}
}
\value{
Speed the player reaches at a given time from zero velocity
}
\description{
Speed over time
}
