% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_distance.R
\name{time_distance}
\alias{time_distance}
\title{Time Distance}
\usage{
time_distance(max_speed, player_tau, distance)
}
\arguments{
\item{max_speed}{Max observed speed of the player}

\item{player_tau}{A unique tau value for each player}

\item{time}{Time}
}
\value{
How long it takes to reach a given distance from zero velocity
}
\description{
Time Distance
}
