% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_speed.R
\name{time_speed}
\alias{time_speed}
\title{Time Speed}
\usage{
time_speed(max_speed, current_speed, player_tau)
}
\arguments{
\item{max_speed}{Max observed speed of the player}

\item{current_speed}{The player's current speed}

\item{player_tau}{A unique tau value for each player}
}
\value{
How long it takes to reach max speed with a starting speed >= 0
}
\description{
Time Speed
}
